css/css-scroll-snap/snap-to-visible-areas.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
:root {
  overflow: scroll;
  scroll-snap-type: both mandatory;
}
div {
  position: absolute;
  margin: 0px;
}

#space {
  width: 2000px;
  height: 2000px;
}

.snap {
  width: 200px;
  height: 200px;
  background-color: blue;
  scroll-snap-align: start;
}

#left-top {
  left: 0px;
  top: 0px;
}

#right-top {
  left: 1000px;
  top: 0px;
}

#left-bottom {
  left: 0px;
  top: 800px;
}

</style>
<div id="space"></div>
<div id="left-top" class="snap"></div>
<div id="right-top" class="snap"></div>
<div id="left-bottom" class="snap"></div>

<script>
test(() => {
  window.scrollTo(0, 0);
  assert_equals(window.scrollX, 0);
  assert_equals(window.scrollY, 0);
  window.scrollTo(400, 0);
  assert_equals(window.scrollX, 1000);
  assert_equals(window.scrollY, 0);
}, 'Only snap to visible area on X axis, even when the non-visible ones are closer');

test(() => {
  window.scrollTo(0, 0);
  assert_equals(window.scrollX, 0);
  assert_equals(window.scrollY, 0);
  window.scrollTo(0, 300);
  assert_equals(window.scrollX, 0);
  assert_equals(window.scrollY, 800);
}, 'Only snap to visible area on Y axis, even when the non-visible ones are closer');
</script>
